Add files via upload
WhatsApp is a great source of data to analyze many patterns and relationships between two or more people chatting personally or even in groups. If you want to know how we can analyze the sentiments of a WhatsApp chat, this article is for you. In this simple project, I will walk you through the task of WhatsApp chat sentiment analysis using Python.